refactor: Place common modal components in single component
