+++ src/routes/popup/announcement.tsx
@@ -1,16 +1,15 @@
-import {
-  ButtonV2,
-  ModalV2,
-  Spacer,
-  Text,
-  type DisplayTheme
-} from "@arconnect/components";
+import { ButtonV2, ModalV2, Spacer } from "@arconnect/components";
 import { ExtensionStorage } from "~utils/storage";
 import { useEffect, useRef, useState } from "react";
 import browser from "webextension-polyfill";
 import aoLogo from "url:/assets/ecosystem/ao-token-logo.png";
-import styled from "styled-components";
 import { useStorage } from "@plasmohq/storage/hook";
+import {
+  ContentWrapper,
+  Content,
+  HeaderText,
+  CenterText
+} from "~components/modals/Components";
 
 export const AnnouncementPopup = ({ isOpen, setOpen }) => {
   const [notifications, setNotifications] = useStorage<boolean>({
@@ -26,11 +25,6 @@ export const AnnouncementPopup = ({ isOpen, setOpen }) => {
       setChecked(notifications);
     }
   }, [notifications]);
-
-  const handleCheckbox = async () => {
-    setChecked((prev) => !prev);
-    setNotifications((prev) => !prev);
-  };
 
   const handleClickOutside = (event) => {
     if (modalRef.current && !modalRef.current.contains(event.target)) {
@@ -87,83 +81,3 @@ export const AnnouncementPopup = ({ isOpen, setOpen }) => {
     </ModalV2>
   );
 };
-
-export const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  flex: none;
-  align-self: stretch;
-  flex-grow: 0;
-`;
-
-export const ContentWrapper = styled.div`
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: space-between;
-`;
-
-const CenterText = styled(Text).attrs({
-  noMargin: true
-})<{ displayTheme?: DisplayTheme }>`
-  width: 245px;
-  text-align: center;
-  color: ${(props) =>
-    props.theme.displayTheme === "light" ? "#191919" : "#FFFFFF"};
-  font-weight: 500;
-  font-size: 11px;
-  line-height: 16px;
-  align-self: stretch;
-  flex: none;
-  flex-grow: 0;
-`;
-
-const Link = styled.u`
-  cursor: pointer;
-`;
-
-const CheckContainer = styled.div`
-  width: 245px;
-  display: flex;
-  flex-direction: row;
-  padding-left: 72px;
-  align-items: center;
-  isolation: isolate;
-  font-weight: 500;
-  font-size: 11px;
-  flex: none;
-  flex-grow: 0;
-  gap: 8px;
-`;
-
-const CheckedSvg = styled.svg`
-  position: absolute;
-  left: calc(50% + 4px / 2 - 113px);
-  width: 18px;
-  height: 18px;
-  cursor: pointer;
-  flex: none;
-  flex-grow: 0;
-  background: #8e7bea;
-  border-radius: 2px;
-`;
-
-const UncheckedSvg = styled.svg`
-  position: absolute;
-  left: calc(50% + 4px / 2 - 113px);
-  width: 18px;
-  height: 18px;
-  cursor: pointer;
-  flex: none;
-  flex-grow: 0;
-`;
-
-const HeaderText = styled(Text)<{ displayTheme?: DisplayTheme }>`
-  font-size: 18px;
-  font-weight: 500;
-  color: ${(props) =>
-    props.theme.displayTheme === "light" ? "#191919" : "#FFFFFF"};
-`;
