+++ src/routes/popup/send/announcement.tsx
@@ -1,14 +1,13 @@
-import {
-  ButtonV2,
-  ModalV2,
-  Spacer,
-  Text,
-  type DisplayTheme
-} from "@arconnect/components";
+import { ButtonV2, ModalV2, Spacer } from "@arconnect/components";
 import { useRef } from "react";
 import browser from "webextension-polyfill";
 import aoLogo from "url:/assets/ecosystem/ao-token-logo.png";
-import styled from "styled-components";
+import {
+  HeaderText,
+  CenterText,
+  Content,
+  ContentWrapper
+} from "~components/modals/Components";
 
 export const AnnouncementPopup = ({ isOpen, setOpen }) => {
   const modalRef = useRef(null);
@@ -60,48 +59,3 @@ export const AnnouncementPopup = ({ isOpen, setOpen }) => {
     </ModalV2>
   );
 };
-
-const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  flex: none;
-  align-self: stretch;
-  flex-grow: 0;
-`;
-
-const ContentWrapper = styled.div`
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: space-between;
-`;
-
-const CenterText = styled(Text).attrs({
-  noMargin: true
-})<{ displayTheme?: DisplayTheme }>`
-  width: 245px;
-  text-align: center;
-  color: ${(props) =>
-    props.theme.displayTheme === "light" ? "#191919" : "#FFFFFF"};
-  font-weight: 500;
-  font-size: 11px;
-  line-height: 16px;
-  align-self: stretch;
-  flex: none;
-  flex-grow: 0;
-
-  a {
-    color: rgb(${(props) => props.theme.theme});
-    text-decoration: none;
-  }
-`;
-
-const HeaderText = styled(Text)<{ displayTheme?: DisplayTheme }>`
-  font-size: 18px;
-  font-weight: 500;
-  color: ${(props) =>
-    props.theme.displayTheme === "light" ? "#191919" : "#FFFFFF"};
-`;
